requirements.txt
argcomplete==1.8.2
asgiref==3.3.1
base58==0.2.4
boto3==1.4.4
botocore==1.5.40
certifi==2020.11.8
cffi==1.14.3
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
dataclasses==0.8
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.43.0
django-rest-auth==0.9.5
django-rest-swagger==2.1.2
django-s3-storage==0.13.4
django-storages==1.10.1
djangorestframework==3.7.7
djongo==1.2.38
dnspython==2.0.0
docutils==0.16
durationpy==0.4
future==0.16.0
futures==3.1.1
hjson==2.0.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.2
kappa==0.6.0
lambda-packages==0.15.0
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
placebo==0.9.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.6.0
python-slugify==1.2.4
python3-openid==3.2.0
pytz==2020.4
PyYAML==3.12
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.1.13
simplejson==3.17.2
six==1.10.0
sqlparse==0.2.4
toml==0.9.2
tqdm==4.11.2
troposphere==2.6.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.22
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.42.0
